Refactor QueryExtensions
