+++ decidim-api/lib/decidim/api/schema.rb
@@ -11,7 +11,7 @@ module Decidim
       max_depth Decidim::Api.schema_max_depth
       max_complexity Decidim::Api.schema_max_complexity
 
-      orphan_types(Api.orphan_types)
+      orphan_types(*(Api.orphan_types + [Decidim::Core::UserType]))
     end
   end
 end
